+++ ec/src/lib.rs
@@ -236,6 +236,7 @@ pub trait AffineCurve:
     + Copy
     + Clone
     + Default
+    + UniformRand
     + Send
     + Sync
     + Hash